backend/requirements.txt
# Model Service Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
Pillow>=10.0.0
llama-cpp-python>=0.2.0
